# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/v1/students/populate:
    # binds a127 app logic to a route
    x-swagger-router-controller: student
    get:
      description: Returns all student records with populate
      # used as the method name of the controller
      operationId: student
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/populate/emp1/findall:
    # binds a127 app logic to a route
    x-swagger-router-controller: emp1
    get:
      description: Returns all employe
      # used as the method name of the controller
      operationId: findAllEmployee
        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/populate/emp1/saveone:
    # binds a127 app logic to a route
    x-swagger-router-controller: emp1
    post:
      description: Returns all employe
      # used as the method name of the controller
      operationId: saveOneEmployee
      parameters:
        # BODY PARAMS
        - name: EmpBody
          in: body
          schema:
            $ref: "#/definitions/EmpBody"
        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/populate/book1/findall:
    # binds a127 app logic to a route
    x-swagger-router-controller: book1
    get:
      description: Returns all book
      # used as the method name of the controller
      operationId: findAllBooks
        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/populate/book1/saveone:
    # binds a127 app logic to a route
    x-swagger-router-controller: book1
    post:
      description: add one book
      # used as the method name of the controller
      operationId: saveOneBook
      parameters:
        # BODY PARAMS
        - name: BookBody
          in: body
          schema:
            $ref: "#/definitions/BookBody"
        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/populate/course1/findall:
    # binds a127 app logic to a route
    x-swagger-router-controller: course1
    get:
      description: Returns all book
      # used as the method name of the controller
      operationId: findAllCourses
        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/populate/course1/saveone:
    # binds a127 app logic to a route
    x-swagger-router-controller: course1
    post:
      description: add one course
      # used as the method name of the controller
      operationId: saveOneCourse
      parameters:
        # BODY PARAMS
        - name: CourseBody
          in: body
          required: true
          schema:
            $ref: "#/definitions/CourseBody"
        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/populate/course1/updateone/{course_id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: course1
    put:
      description: update one course
      # used as the method name of the controller
      operationId: updateOneCourse
      parameters:
        # MANDATORY HEADERS
        - name: course_id
          in: path
          type: string
          required: true
        # BODY PARAMS
        - name: CourseBody
          in: body
          required: true
          schema:
            $ref: "#/definitions/CourseBody1"
        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/populate/customer1/findall:
    # binds a127 app logic to a route
    x-swagger-router-controller: customer1
    get:
      description: Returns all customers
      # used as the method name of the controller
      operationId: findAllCustomers
        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/populate/customer1/saveone:
    # binds a127 app logic to a route
    x-swagger-router-controller: customer1
    post:
      description: add one customer
      # used as the method name of the controller
      operationId: saveOneCustomer
      parameters:
        # BODY PARAMS
        - name: CustomerBody
          in: body
          required: true
          schema:
            $ref: "#/definitions/CustomerBody"
        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/populate/customer1/updateone/{customer_id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: customer1
    put:
      description: update one customer
      # used as the method name of the controller
      operationId: updateOneCustomer
      parameters:
        # MANDATORY HEADERS
        - name: customer_id
          in: path
          type: string
          required: true
        # BODY PARAMS
        - name: CustomerBody
          in: body
          required: true
          schema:
            $ref: "#/definitions/CustomerBody1"
        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  EmpBody:
    type: object
    properties:
      eid:
        type: number
      ename:
        type: string
      salary:
        type: number
  BookBody:
    type: object
    properties:
      bid:
        type: number
      bname:
        type: string
      price:
        type: number
      type:
        type: string
  CourseBody:
    type: object
    properties:
      course_id:
        type: number
      course_name:
        type: string
      fee:
        type: number
  CourseBody1:
    type: object
    properties:
      book_id:
        type: string
  CustomerBody:
    type: object
    properties:
      cid:
        type: number
      cname:
        type: string
      age:
        type: number
  CustomerBody1:
    type: object
    properties:
      course_id:
        type: string
      emp_id:
        type: array
        items: {
          type: string
        }